refactor: 마이페이지 UI 구조 개선 및 뱃지 목록 그리드 전환
* **refactor: `MyScreen` 레이아웃을 `LazyVerticalGrid` 기반으로 변경**
    * 기존 `Column` 내부에 고정되었던 UI 구조를 2열 `LazyVerticalGrid`로 전환하여 뱃지 목록을 그리드 형태로 표시하도록 개선했습니다.
    * `ProfileSection`과 `BadgeListTitle`이 그리드 내에서 전체 너비를 차지하도록 `GridItemSpan(maxLineSpan)`을 적용했습니다.
    * 뱃지 목록의 각 아이템을 처리하는 `BadgeGridItem` 컴포저블을 추가했습니다.

* **refactor: 뱃지 섹션 구성 방식 변경 및 파일 정리**
    * 가로 스크롤 방식(`LazyRow`)을 사용하던 `BadgeSection.kt` 파일을 삭제하고 해당 로직을 그리드 구조에 맞게 `MyScreen`으로 통합했습니다.
    * 뱃지 타이틀 클릭 시 리플 효과가 없는 `noRippleClickable`을 적용하여 인터랙션을 조정했습니다.

* **cleanup: 불필요한 설정 및 어노테이션 제거**
    * 더 이상 사용되지 않는 `ExperimentalMaterial3Api` 어노테이션을 삭제했습니다.
    * 그리드 시스템에 맞춰 패딩 및 간격(`Arrangement.spacedBy`) 설정을 최적화했습니다.
