+++ Prezel/feature/my/impl/src/main/java/com/team/prezel/feature/my/impl/MyScreen.kt
@@ -1,24 +1,36 @@
 package com.team.prezel.feature.my.impl
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.PaddingValues
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
-import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.foundation.lazy.grid.GridCells
+import androidx.compose.foundation.lazy.grid.GridItemSpan
+import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
+import androidx.compose.foundation.lazy.grid.items
+import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalResources
+import androidx.compose.ui.res.painterResource
 import androidx.hilt.lifecycle.viewmodel.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import com.team.prezel.core.designsystem.component.feedback.snackbar.showPrezelSnackbar
+import com.team.prezel.core.designsystem.component.list.PrezelList
+import com.team.prezel.core.designsystem.component.list.PrezelListSize
+import com.team.prezel.core.designsystem.icon.PrezelIcons
 import com.team.prezel.core.designsystem.preview.BasicPreview
 import com.team.prezel.core.designsystem.theme.PrezelTheme
+import com.team.prezel.core.ui.component.PrezelBadge
 import com.team.prezel.core.ui.state.LocalSnackbarHostState
-import com.team.prezel.feature.my.impl.component.BadgeSection
+import com.team.prezel.core.ui.util.noRippleClickable
 import com.team.prezel.feature.my.impl.component.MyTopAppBar
 import com.team.prezel.feature.my.impl.component.ProfileSection
 import com.team.prezel.feature.my.impl.contract.MyUiEffect
@@ -28,7 +40,6 @@ import com.team.prezel.feature.my.impl.model.BadgeUiModel
 import com.team.prezel.feature.my.impl.model.MyUiMessage
 import kotlinx.collections.immutable.toImmutableList
 
-@OptIn(ExperimentalMaterial3Api::class)
 @Composable
 internal fun MyScreen(
     navigateToEditProfile: () -> Unit,
@@ -76,26 +87,83 @@ private fun MyScreen(
     modifier: Modifier = Modifier,
 ) {
     Column(
-        modifier = modifier.fillMaxSize().background(PrezelTheme.colors.bgRegular),
+        modifier = modifier
+            .fillMaxSize()
+            .background(PrezelTheme.colors.bgRegular),
         horizontalAlignment = Alignment.CenterHorizontally,
     ) {
         MyTopAppBar(onClickSetting = onClickSetting)
 
-        Spacer(modifier = Modifier.height(PrezelTheme.spacing.V16))
+        LazyVerticalGrid(
+            columns = GridCells.Fixed(2),
+            modifier = Modifier.fillMaxWidth(),
+            contentPadding = PaddingValues(
+                vertical = PrezelTheme.spacing.V16,
+                horizontal = PrezelTheme.spacing.V20,
+            ),
+            verticalArrangement = Arrangement.spacedBy(PrezelTheme.spacing.V16),
+            horizontalArrangement = Arrangement.spacedBy(PrezelTheme.spacing.V16),
+        ) {
+            item(span = { GridItemSpan(maxLineSpan) }) {
+                ProfileSection(
+                    nickname = uiState.nickname,
+                    profileImageUrl = uiState.profileImageUrl,
+                    onClickEditProfile = onClickEditProfile,
+                    modifier = Modifier.fillMaxWidth(),
+                )
+            }
 
-        ProfileSection(
-            nickname = uiState.nickname,
-            profileImageUrl = uiState.profileImageUrl,
-            onClickEditProfile = onClickEditProfile,
-        )
+            item(span = { GridItemSpan(maxLineSpan) }) {
+                Spacer(modifier = Modifier.height(PrezelTheme.spacing.V16))
+            }
 
-        Spacer(modifier = Modifier.height(PrezelTheme.spacing.V32))
+            item(span = { GridItemSpan(maxLineSpan) }) {
+                BadgeListTitle(
+                    modifier = Modifier
+                        .fillMaxWidth()
+                        .noRippleClickable(onClick = navigateToBadge),
+                )
+            }
 
-        BadgeSection(
-            badges = uiState.badges,
-            onClickBadge = navigateToBadge,
-        )
+            items(
+                items = uiState.badges,
+                key = { badge -> badge.code },
+            ) { badge ->
+                BadgeGridItem(badge = badge)
+            }
+        }
     }
+}
+
+@Composable
+private fun BadgeListTitle(modifier: Modifier = Modifier) {
+    PrezelList(
+        title = "나의 뱃지",
+        titleTextColor = PrezelTheme.colors.textLarge,
+        size = PrezelListSize.REGULAR,
+        nested = true,
+        modifier = modifier.fillMaxWidth(),
+        trailingContent = {
+            Icon(
+                painter = painterResource(PrezelIcons.ChevronRight),
+                contentDescription = "뱃지",
+                tint = PrezelTheme.colors.iconRegular,
+            )
+        },
+    )
+}
+
+@Composable
+private fun BadgeGridItem(
+    badge: BadgeUiModel,
+    modifier: Modifier = Modifier,
+) {
+    PrezelBadge(
+        title = badge.title,
+        url = badge.imageUrl,
+        isAchieved = badge.isAchieved,
+        modifier = modifier.fillMaxWidth(),
+    )
 }
 
 @BasicPreview
